Sign and verify should follow default policies and have consistent messages
