+++ src/NuGet.Core/NuGet.Packaging/Signing/Authoring/ITimestampProvider.cs
@@ -9,7 +9,7 @@ namespace NuGet.Packaging.Signing
 {
     public interface ITimestampProvider
     {
-        // Sign and timestamp a file.
+        // Add timestamp to signature
         Task<Signature> TimestampSignatureAsync(TimestampRequest request, ILogger logger, CancellationToken token);
     }
 }
