+++ src/NuGet.Core/NuGet.Packaging/Signing/NuGetChainBuildingRequestType.cs
@@ -1,0 +1,12 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+namespace NuGet.Packaging.Signing
+{
+    public enum NuGetChainBuildingRequestType
+    {
+        Unknown         = 0,
+        Signing         = 1,
+        Verification    = 2
+    }
+}
