+++ src/NuGet.Core/NuGet.Packaging/Signing/Utility/AttributeUtility.cs
@@ -154,7 +154,7 @@ namespace NuGet.Packaging.Signing
         {
             return IsValidSigningCertificateV2(
                 signatureCertificate: signatureCertificate,
-                chain: SigningUtility.GetCertificateChain(chain),
+                chain: SigningUtility.CertificateChainToList(chain),
                 signingCertV2Attribute: signingCertV2Attribute,
                 signingSpecifications: signingSpecifications);
         }
